shell
A core mechanic of the module system is that option values can depend on
arbitrary other option values.

To demonstrate that, I've introduced a new option `shellInit` that contains the echo command we want to run,
while the activation script now starts a bash shell with the `shellInit`
as the `--rcfile` argument.

We're making use of the magical built-in Nix interpolation syntax here.
Due to it, if we build the activation script, Nix will automatically:
- Realise that it depends on `shellInit`, so that needs to be built
  first
- Insert the resulting store path into the interpolation when building
  the activation script
- Register the `shellInit` as a dependency of the activation script