python
Next, I want to introduce the concept of abstracting configuration with
the module system.

Let's say that many users need to use Python,
including the interpreter, a dependency manager, perhaps also a
formatter. While every user could specify all the packages they need
themselves, it would be really convenient if we could _abstract_ away
the idea of Python development. And indeed, we can do that, which is
what the new `python.nix` module here shows. Note that we could also use
the same module for that due to how `imports` works, but it's cleaner if we use a separate one.

The `python` module here defines a boolean `python` option, which when
enabled adds two default python-related packages.
This new option allows the user to just set `python = true;` to get
these packages, no need to list them manually!

(try it out)

While this example is very basic for now, we'll take this to the next
level soon.

FAQ:
- `imports` vs conditional definitions?
  - Can't unimport modules again
  - Can have more complex conditions
  - Might want to define an option even if it's not being used right
    now, in case it's later