home
Now let's go back to our base module and make a small but mighty change:
Instead of having the activation script start a new shell with the
initial script, let's install the script to the users standard shell
initialisation file!

Now the activation script will exit, but have a lasting effect, let's
try it out:

(build & activate)
(run a new `zsh`, demo that the packages are there)

As we can see, whenever we start our shell now, we get the greeting, and
we get all the packages available on `PATH`, we've now effectively "installed" these
packages!

With this small change, we've now completely changed the underlying model: This is
now essentially a user dotfile management tool. In the Nix community, the
home-manager project is the full-fledged equivalent now.

We're also defining the _class_ of the module here, to emphasize that
this is a user-level module, which is a bit of foreshadowing for the
following slide.