packages
Until now we haven't seen much that can be configured, let's change
that by adding a `packages` option to specify packages to add to the `PATH` environment
variable, making their binaries available to run inside the shell.

We're seeing some new things here, but they're not that important for
the story:
- A new type, `listOf`, to specify a list of packages
- A default option value
- The `makeBinPath` function to turn a list of packages into a string
  suitable for `PATH`

More importantly, let's consider the use case of providing this module
for end users to specify packages. While we could just send this file to
the users and tell them what they need to edit, that's very inconvenient
for the user and for updates. To address this, we'll make use of another
core mechanic of the module system: `imports`.

(open custom.nix)

`imports` are essentially just a way of splitting modules into smaller
more specialised parts, that's what makes the module system _modular_. In our case, we
have a custom.nix module for users to write, which imports a base.nix
module distributed by developers.

To the module system, there is no difference between the two modules,
it evaluates the same whether it's a single combined module or the two
separate ones.

Moving on, we're now adding some packages, `fortune` and `cowsay`. These
come from Nixpkgs, the package collection for Nix, containing about
100000 packages in total. Let's try it out

(activate)
(run `fortune -s computers | cowsay`)

Note how these package are only available inside the shell, and they
don't conflict with any packages already on the system due to how Nix
only stores the packages in the Nix store with its hashes.