neo
Now that we know it works for one user, let's actually define another
user! Here we define the user neo and add the cmatrix package to the
environment.

(build and run, ssh to the user, run cmatrix)

At this point, we've seen the core mechanics of the module system:
- How we can modularise with multiple modules and `imports`
- How we can abstract with options
- How we can nest independent module evaluations within each other

While the module system does a whole bunch more, for the rest of this
talk, for however long I still have, I'll focus on expanding what we
have, to emphasize how powerful these core mechanics are.