motd
With our activation script now being able to configure an entire
host, let's also add a host-level option that's not specific to users.

Here we define the `etc` option, which allows declaring contents of
entries in the `/etc` directory.

We'll use it to set the message of the day, which gets displayed for all
users when we log into the machine.

(build and activate, ssh to any user)

At this point, we've essentially built a very minimal and
feature-incomplete version of NixOS,
an entire linux distribution built with the module system.