another host
Now let's make use of this new level by declaring the configuration of
some more hosts, `modus` and `create`.

We can now easily reach for any of the previous abstractions,
independently of the other hosts:
- On the `modus` host we specify that `alex` is a developer, which
  automatically creates the user, and adds multiple python tools to
  their environment
- On the `tweag` host, we directly specify that the root user should
  have `htop` available

To make this work with non-local hosts however,
we need one small additional change: This `nix-copy-closure` command

The standard is for Nix to store its paths on the local host, so if we
just tried to run the activation script on another host, it's not going to be there.

nix-copy-closure takes care of not only copying the activation script,
but also all of its runtime dependencies, such as htop in this case.