activationScript
Let's make it do something shall we?

The core mechanic of the module system is a namespace of _options_.

Options are a lot like fields in structs: They have a name and a type.

Modules have two main sections to interact with the option namespace:
- The `options` section declares options
- The `config` section gives them values

While we can define any options we want, options can't _do_ anything by
themselves, due to the fact that Nix is a _purely_ functional language,
there are no side effects.

Because of that, virtually all use cases of the module system use the
concept of an activation script, which we've declared as a first option
here. Its type is a _package_ which is Nix's built-in concept of a buildable directory.
As the value, we make use of a utility to write a shell script to echo a string.

We can now use the Nix CLI to _evaluate_ this option and _build_ the package to give us
this directory with the shell script (though I've hidden some detail for
brevity):

> nix-build
> ls

As you can see, Nix built the package, created the directory in the Nix
store, and gave us a local `result` symlink pointing to it.

We can now run the activation script:

> result/bin/activate