empty module
Today I'd like to show you a glimpse of the power of the Nixpkgs module
system. The module system is at the core of NixOS but also many other
projects such as home-manager, flake-parts, devenv and more.

Over the course of this presentation, I'll slowly introduce more and
more core mechanism and features of the module system.

For now the main thing to know is
that the module system is a DSL written in Nix, and it's based around the concept of _modules_.

Here we have almost the simplest module you can write: A module that
doesn't define anything. It does have _arguments_ at the top, which
we'll use later. But for now they're just witness to
the fact that modules are just Nix _functions_, though that's not very
important to know for this presentation.